% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces_concepts_from_qcew.R
\name{generate_ces_from_qcew}
\alias{generate_ces_from_qcew}
\title{Generate CES concepts from QCEW data}
\usage{
generate_ces_from_qcew(qcew_area_code, year_start)
}
\arguments{
\item{qcew_area_code}{A character string representing the QCEW area code.}

\item{year_start}{A numeric four-digit year indicating the start of the data retrieval.}
}
\value{
A tibble containing aggregated employment data.
}
\description{
Downloads national CES industry data and aggregates QCEW data into CES concepts.
}
